% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gff.R
\name{ancestor}
\alias{ancestor}
\title{Compute the ancestor of a gff feature}
\usage{
ancestor(.data, kid = feature_id, parent = parent_id)
}
\description{
Bubble up the feature_id -> parent_id associations until the top-level ID is
found. This is the ancestor_id, shared by all associated features.
}
