% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{as.multiPhylo.list}
\alias{as.multiPhylo.list}
\title{Convert a List into multiPhylo Object}
\usage{
as.multiPhylo.list(x, ...)
}
\arguments{
\item{x}{list of phylo objects}
}
\value{
multiPhylo object
}
\description{
Turn a list of phylo objects into a multiPhylo object by adding "multiPhylo"
as class attribute. Method for generic function `phytools::as.multiPhylo`.
}
