% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{distance}
\alias{distance}
\title{Compute great circle distances}
\usage{
distance(
  .data,
  lat = lat,
  long = long,
  lat_from = NULL,
  long_from = NULL,
  dist_col = "dist",
  along = TRUE,
  ...
)
}
\arguments{
\item{.data}{a tbl with coordinates in degrees}

\item{lat, long}{latitude and longitude columns (unquoted)}

\item{lat_from, long_from}{optional reference point coordinates}

\item{dist_col}{name of the output column}

\item{along}{if TRUE return cumulative distances along points}

\item{...}{arguments passed on to distVincentyEllipsoid}
}
\description{
Wraps \code{\link[geosphere]{distVincentyEllipsoid}} in tidy semantics to
compute and add the shortest distance between to points on a globe to a
table of coordinates.
}
\examples{
stations <- tibble(
  lat=c(50,55,60),
  long=c(180,180,180)
)

# cumulative distance along transect stations
stations \%>\% distance

# distances to almost Greenwich
stations \%>\%
  distance(lat_from=0, long_from=0, along=FALSE)
}
