% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-vars.R, R/todo.R
\name{enable_TODO}
\alias{enable_TODO}
\alias{disable_TODO}
\title{Activate TODO() messages}
\usage{
enable_TODO()

enable_TODO()

disable_TODO()
}
\description{
`TODO("msg")` are suppressed by default. `enable_todos()` enables them for
following code. `deactivate_todos()` disables them again.

`TODO("msg")` are suppressed by default. `enable_todos()` enables them for
following code. `deactivate_todos()` disables them again.
}
