% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{flip_descendants}
\alias{flip_descendants}
\title{Flip Descendant Clades}
\usage{
flip_descendants(tree_view = NULL, node)
}
\arguments{
\item{tree_view}{tree view}

\item{node}{node}
}
\value{
ggtree object
}
\description{
Flip two clades descending from `node` in a ggtree. Throw an error for
ambigious cases.
}
