% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{get_children}
\alias{get_children}
\title{Get child nodes}
\usage{
get_children(tree, node)
}
\arguments{
\item{tree}{phylo object}

\item{node}{parent node}
}
\description{
Get the immediate descendants for a given node.
}
