% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{get_mrca}
\alias{get_mrca}
\title{Most Recent Common Ancestor}
\usage{
get_mrca(tree, tips)
}
\arguments{
\item{tree}{phylo object}

\item{tips}{vector of tips (or nodes)}
}
\description{
Get the most recent common ancester for a set of tips. Wraps `ape::getMRCA`
and also works if only a subset of tips is actually present in the tree
}
\examples{
# "t6" is not present in tree
get_mrca(ape::rtree(5), tips=c("t4","t5","t6"))
}
