% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{ggplot default color vector}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colors}
}
\description{
Emulate ggplot default colors for a given number of categories. Credit to
John Colby. https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
