% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrx.R
\name{index_by}
\alias{index_by}
\title{Index by keys preserving order}
\usage{
index_by(.data, ...)
}
\arguments{
\item{keys}{in \code{\link[dplyr]{select}} semantics}
}
\description{
Add indices by key columns while preserving order according to the first occurence
}
\examples{
tibble(x=c(1,1,1,2), y=c("B", "A", "B", "B"), z="foo") \%>\%
  index_by(x,y)
}
