% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{mount_mit}
\alias{mount_mit}
\title{Mount MIT remote drives}
\usage{
mount_mit()
}
\description{
Mount MIT remote drives (/nobackup1) hosted on engaging using my zsh-alias
}
\examples{
mount_mit()
}
