% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{outgroup_root}
\alias{outgroup_root}
\title{Reroot Tree by Outgroup}
\usage{
outgroup_root(tree, tips)
}
\arguments{
\item{tree}{phylo object}

\item{tips}{outgroup tip label}
}
\value{
phylo object
}
\description{
Reroot a tree based on set of outgroup tip labels. Sets the root to the
outgroup MRCA.
}
