% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools_update.R
\name{pkg_update}
\alias{pkg_update}
\title{Document and reload dev package}
\usage{
pkg_update(pgk = here::here())
}
\arguments{
\item{pkg}{package to update, defaults to `here::here()`}
}
\description{
Document and reload dev package
}
