% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blast.R
\name{read_blast}
\alias{read_blast}
\title{Read a BLAST .tsv files}
\usage{
read_blast(file, col_spec = "std")
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
be automatically uncompressed. Files starting with \verb{http://},
\verb{https://}, \verb{ftp://}, or \verb{ftps://} will be automatically
downloaded. Remote gz files can also be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. It must contain at
least one new line to be recognised as data (instead of a path) or be a
vector of greater than length 1.

Using a value of \code{\link[readr:clipboard]{clipboard()}} will read from the system clipboard.}

\item{col_spec}{BLAST custom column specifiers}
}
\value{
tibble
}
\description{
Read a BLAST output in `-outfmt 6/7` format (o6/o7). If BLAST was run
with custom column specifiers, those have to be passed to the call.
}
\details{
Uses "# " to identify and ignore comment lines in o7. NOTE that any line
containing this string anywhere, not limited to the line start will be
trimmed after the "# ".
}
