% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fai.R
\name{read_fai}
\alias{read_fai}
\title{Read fasta index}
\usage{
read_fai(
  file,
  col_names = c("seq_id", "seq_desc", "length"),
  keep_index = FALSE
)
}
\arguments{
\item{file}{.fai/.seqkit.fai (plain/pipe), .fa (plain)}

\item{col_names}{custom column names}

\item{keep_index}{do not ignore index columns with length offsets etc.}
}
\value{
tibble
}
\description{
Read seq_ids and lengths from a samtools/seqkit faidx file or from a fasta
directly (requires seqkit in path). Seqkit is preferred because it allows
reading the full header including seq description.
}
\details{
Note, only .fai connections can be pipes, .fa have to be plain text files.
}
\examples{
read_fai("foo.fa.fai")  # samtools style index, no description
read_fai("foo.fa.seqkit.fai") # seqkit style with description
# custom col_names and keep index info
read_fai("foo.fa.fai", keep_index=T, col_names=c("contig_id", "contig_desc", "len"))
}
