% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gff.R
\name{read_gff}
\alias{read_gff}
\title{Read a .gff file}
\usage{
read_gff(gff_file, sources = NULL, types = NULL)
}
\arguments{
\item{gff_file}{to read}

\item{sources}{only return features from these sources}

\item{types}{only return features of these types, e.g. gene, CDS, ...}
}
\value{
tibble
}
\description{
Uses `rtracklayer::import`.
}
