% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hmmer.R
\name{read_hmmer}
\alias{read_hmmer}
\title{Read a HMMER .tbl/tsv file}
\usage{
read_hmmer(file)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
be automatically uncompressed. Files starting with \verb{http://},
\verb{https://}, \verb{ftp://}, or \verb{ftps://} will be automatically
downloaded. Remote gz files can also be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. It must contain at
least one new line to be recognised as data (instead of a path) or be a
vector of greater than length 1.

Using a value of \code{\link[readr:clipboard]{clipboard()}} will read from the system clipboard.}
}
\value{
tibble
}
\description{
Read a HMMER tblout file from a protein search. The tbl file is converted to
tab-separated file using `hmmer-tbl2tsv`. Can also read tsv generated by
`hmmer-tbl2tsv` directly.
}
