% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{reroot_by_outgroup}
\alias{reroot_by_outgroup}
\title{Reroot Tree by Outgroup}
\usage{
reroot_by_outgroup(tree, tips)
}
\arguments{
\item{tree}{phylo object}

\item{tips}{outgroup tip labels}
}
\value{
phylo object
}
\description{
Reroot a tree based on set of outgroup tip labels. Sets the root to the
outgroup MRCA.
}
