% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtree.R
\name{scale_y_tree}
\alias{scale_y_tree}
\title{Continious y-scale with fixed expansion space}
\usage{
scale_y_tree(expand = expand_scale(0, 0.6), ...)
}
\arguments{
\item{expand}{Vector of range expansion constants used to add some
padding around the data, to ensure that they are placed some distance
away from the axes. Use the convenience function \code{\link[=expand_scale]{expand_scale()}}
to generate the values for the \code{expand} argument. The defaults are to
expand the scale by 5\% on each side for continuous variables, and by
0.6 units on each side for discrete variables.}

\item{...}{Other arguments passed on to \code{scale_(x|y)_continuous}}
}
\description{
Overwrites the default expand for continuous scales
}
