% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_info_diff.R
\name{session_info_diff}
\alias{session_info_diff}
\title{Compare sessionInfo()s}
\usage{
session_info_diff(x = readLines("clipboard"), y = sessionInfo())
}
\description{
Compare a sessionInfo string for example from a github issue to the current
session and show the differences. Just copy the query session info from the
issue post into the clipboard and call `session_info_diff` directly.
}
