% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class.R
\name{set_class}
\alias{set_class}
\title{set class}
\usage{
set_class(x, class, add = c("overwrite", "prepend", "append"))
}
\arguments{
\item{x}{Object to assign new class to.}

\item{class}{Class value to assign to x}
}
\value{
Object x as class value.
}
\description{
Set class of an object. Optionally append or prepend to exiting class
attributes.
}
