% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringrx.R
\name{str_which_unique}
\alias{str_which_unique}
\alias{str_subset_unique}
\title{Keep strings uniquely matching a pattern, or find positions.}
\usage{
str_which_unique(strings, patterns, require_match = FALSE)

str_subset_unique(strings, patterns, require_match = FALSE)
}
\arguments{
\item{strings}{a vector of characters}

\item{patterns}{a vector of patterns}

\item{require_match}{for each pattern or throw an error}
}
\description{
These are wrappers for `stringr::str_subset` and `stringr::str_which` that
facilitate all-vs-all matching of multiple strings against multiple patterns,
and require matches to be unique across patterns.
}
