% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R
\name{TODO}
\alias{TODO}
\title{Add TODO statements. See `enable/distable_TODO` for control of when to print
them.}
\usage{
TODO(desc, file = stderr())
}
\arguments{
\item{desc}{of what to do}
}
\description{
Add TODO statements. See `enable/distable_TODO` for control of when to print
them.
}
