% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtree.R
\name{tree_y}
\alias{tree_y}
\title{Get tree-based y-coordinates}
\usage{
tree_y(ggtree, data)
}
\arguments{
\item{ggtree}{a ggtree plot object}

\item{data}{a tibble with a `label` column}
}
\description{
Matches a ggtree and a tibble by shared labels, and returns the respective
y-coordinates for the data.
}
