% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtree.R
\name{tree_ylim}
\alias{tree_ylim}
\title{Get the range of the ggtree y-axis data}
\usage{
tree_ylim(ggtree)
}
\arguments{
\item{ggtree}{a ggtree plot object}
}
\description{
Get the range of the ggtree y-axis data
}
